requirements.txt
boto3==1.15.0
awscli-local==0.20
psycopg2-binary==2.9.5